pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "file-agent-ai"
version = "1.7"
description = "An ai file organizer that can create and move files around!"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "AGPL-3.0"}
authors = [
    {name = "Lukas Reis"}
]
keywords = ["ai", "file-organizer", "automation", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "openai",
    "watchdog",
    "python-dotenv",
]

[project.scripts]
file-agent = "file_agent.agent:main"

[project.urls]
Homepage = "https://github.com/maker-lukas/file-agent"

[tool.setuptools.packages]
find = {include = ["file_agent*"]}

[tool.setuptools.package-data]
file_agent = ["instructions.txt", "*.png"]